requirements.txt
blinker==1.6.2
click==8.1.6
colorama==0.4.6
Flask==2.3.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
Werkzeug==2.3.6
requests==2.31.0
numpy==1.26.4
python-dotenv==1.0.1
flask-crontab==0.1.2
psycopg2-binary==2.9.9
